continue tickets mb finished
almost fin

+++ TicketCode/src/WalkUpTicket.java
@@ -2,7 +2,7 @@ public class WalkUpTicket extends Tickets {
 	public WalkUpTicket(int n){
 		super(n);
 	}
-	public int getPrice(){
-		return 50;
+	public double getPrice(){
+		return 50.0;
 	}
 }
